--- a/memiavl/layout_native.go
+++ b/memiavl/layout_native.go
@@ -43,6 +43,10 @@ func (node *nodeLayout) Height() uint8 {
 	return uint8(node.data[0])
 }
 
+func (node NodeLayout) PreTrees() uint8 {
+	return uint8(node.data[0] >> 8)
+}
+
 func (node *nodeLayout) Version() uint32 {
 	return node.data[1]
 }
@@ -51,7 +55,7 @@ func (node *nodeLayout) Size() uint32 {
 	return node.data[2]
 }
 
-func (node *nodeLayout) KeyNode() uint32 {
+func (node *nodeLayout) KeyLeaf() uint32 {
 	return node.data[3]
 }
 
@@ -61,4 +65,52 @@ func (node *nodeLayout) KeyOffset() uint64 {
 
 func (node *nodeLayout) Hash() []byte {
 	return node.hash[:]
+}
+
+type LeafLayout = *leafLayout
+
+// Nodes is a continuously stored IAVL nodes
+type Leaves struct {
+	leaves []leafLayout
+}
+
+func NewLeaves(buf []byte) (Leaves, error) {
+	// check alignment and size of the buffer
+	p := unsafe.Pointer(unsafe.SliceData(buf))
+	if uintptr(p)%unsafe.Alignof(leafLayout{}) != 0 {
+		return Leaves{}, errors.New("input buffer is not aligned")
+	}
+	size := int(unsafe.Sizeof(leafLayout{}))
+	if len(buf)%size != 0 {
+		return Leaves{}, errors.New("input buffer length is not correct")
+	}
+	leaves := unsafe.Slice((*leafLayout)(p), len(buf)/size)
+	return Leaves{leaves}, nil
+}
+
+func (leaves Leaves) Leaf(i uint32) LeafLayout {
+	return &leaves.leaves[i]
+}
+
+type leafLayout struct {
+	version   uint32
+	keyLen    uint32
+	keyOffset uint64
+	hash      [32]byte
+}
+
+func (leaf *leafLayout) Version() uint32 {
+	return leaf.version
+}
+
+func (leaf *leafLayout) KeyLength() uint32 {
+	return leaf.keyLen
+}
+
+func (leaf *leafLayout) KeyOffset() uint64 {
+	return leaf.keyOffset
+}
+
+func (leaf *leafLayout) Hash() []byte {
+	return leaf.hash[:]
 }