Closes: #1016 

`20%` improvement in `RandomGet` benchmark.

```
$ go test -run=^$ -bench="RandomGet/memiavl-disk" -benchmem ./ -count=4 -tags nativebyteorder
```
```
$ benchstat /tmp/before.txt /tmp/after.txt
goos: darwin
goarch: amd64
pkg: github.com/crypto-org-chain/cronos/memiavl
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                          │ /tmp/before.txt │           /tmp/after.txt            │
                          │     sec/op      │    sec/op     vs base               │
RandomGet/memiavl-disk-12      175.8n ± ∞ ¹   138.2n ± ∞ ¹  -21.37% (p=0.029 n=4)
¹ need >= 6 samples for confidence interval at level 0.95

                          │ /tmp/before.txt │         /tmp/after.txt         │
                          │      B/op       │    B/op      vs base           │
RandomGet/memiavl-disk-12       0.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=4) ²
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal

                          │ /tmp/before.txt │         /tmp/after.txt         │
                          │    allocs/op    │  allocs/op   vs base           │
RandomGet/memiavl-disk-12       0.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=4) ²
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
```

👮🏻👮🏻👮🏻 !!!! REFERENCE THE PROBLEM YOUR ARE SOLVING IN THE PR TITLE AND DESCRIBE YOUR SOLUTION HERE !!!! DO NOT FORGET !!!! 👮🏻👮🏻👮🏻


# PR Checklist:

- [ ] Have you read the [CONTRIBUTING.md](https://github.com/crypto-org-chain/chain-main/blob/master/CONTRIBUTING.md)?
- [ ] Does your PR follow the [C4 patch requirements](https://rfc.zeromq.org/spec:42/C4/#23-patch-requirements)?
- [ ] Have you rebased your work on top of the latest master? 
- [ ] Have you checked your code compiles? (`make`)
- [ ] Have you included tests for any non-trivial functionality?
- [ ] Have you checked your code passes the unit tests? (`make test`)
- [ ] Have you checked your code formatting is correct? (`go fmt`)
- [ ] Have you checked your basic code style is fine? (`golangci-lint run`)
- [ ] If you added any dependencies, have you checked they do not contain any known vulnerabilities? (`go list -json -m all | nancy sleuth`)
- [ ] If your changes affect the client infrastructure, have you run the integration test?
- [ ] If your changes affect public APIs, does your PR follow the [C4 evolution of public contracts](https://rfc.zeromq.org/spec:42/C4/#26-evolution-of-public-contracts)?
- [ ] If your code changes public APIs, have you incremented the crate version numbers and documented your changes in the [CHANGELOG.md](https://github.com/crypto-org-chain/chain-main/blob/master/CHANGELOG.md)?
- [ ] If you are contributing for the first time, please read the agreement in [CONTRIBUTING.md](https://github.com/crypto-org-chain/chain-main/blob/master/CONTRIBUTING.md) now and add a comment to this pull request stating that your PR is in accordance with the [Developer's Certificate of Origin](https://github.com/crypto-org-chain/chain-main/blob/master/CONTRIBUTING.md#developer-certificate-of-originn).

Thank you for your code, it's appreciated! :)